refactor(backend): 単一実装の Service interface 10 対を具象クラスへ集約（C3）
「One adapter means a hypothetical seam」— 13 対中 10 対は実装が 1 つで、
テストも interface ではなく Impl を直接 @InjectMocks しており、seam として
機能していなかった。Impl を interface 名に改名し、interface ファイルを削除。

保留（seam が実在する 3 つ）:
- SystemConfigService（MaintenanceModeInterceptor / MailService が跨モジュール mock）
- MailService（TenantCreatedListener が mock）
- FileStorageService（リモートストレージの第二 adapter が予見される）

rules/backend.md にモジュール規約として明文化。挙動変更なし。
