+++ backend/docs/modulith/components.puml
@@ -23,29 +23,29 @@ Container_Boundary("Application.Application_boundary", "Application", $tags="") 
   Component(Application.Application.Tenant, "Tenant", $techn="Module", $descr="", $tags="", $link="")
 }
 
-Rel(Application.Application.Customer, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
-Rel(Application.Application.Customer, Application.Application.Tenant, "uses", $techn="", $tags="", $link="")
-Rel(Application.Application.Auth, Application.Application.User, "uses", $techn="", $tags="", $link="")
-Rel(Application.Application.Auth, Application.Application.Tenant, "uses", $techn="", $tags="", $link="")
-Rel(Application.Application.Auth, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
-Rel(Application.Application.Cast, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
-Rel(Application.Application.Cast, Application.Application.Tenant, "uses", $techn="", $tags="", $link="")
-Rel(Application.Application.Menu, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
-Rel(Application.Application.Storage, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
-Rel(Application.Application.Storeprofile, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
-Rel(Application.Application.Notification, Application.Application.Tenant, "listens to", $techn="", $tags="", $link="")
-Rel(Application.Application.Notification, Application.Application.Settings, "uses", $techn="", $tags="", $link="")
-Rel(Application.Application.Notification, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
 Rel(Application.Application.Order, Application.Application.Customer, "uses", $techn="", $tags="", $link="")
 Rel(Application.Application.Order, Application.Application.Cast, "uses", $techn="", $tags="", $link="")
 Rel(Application.Application.Order, Application.Application.User, "uses", $techn="", $tags="", $link="")
 Rel(Application.Application.Order, Application.Application.Tenant, "uses", $techn="", $tags="", $link="")
 Rel(Application.Application.Order, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
+Rel(Application.Application.Menu, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
+Rel(Application.Application.Auth, Application.Application.User, "uses", $techn="", $tags="", $link="")
+Rel(Application.Application.Auth, Application.Application.Tenant, "uses", $techn="", $tags="", $link="")
+Rel(Application.Application.Auth, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
+Rel(Application.Application.Storage, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
+Rel(Application.Application.Settings, Application.Application.Shared, "depends on", $techn="", $tags="", $link="")
+Rel(Application.Application.Cast, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
+Rel(Application.Application.Cast, Application.Application.Tenant, "uses", $techn="", $tags="", $link="")
+Rel(Application.Application.User, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
+Rel(Application.Application.Notification, Application.Application.Tenant, "listens to", $techn="", $tags="", $link="")
+Rel(Application.Application.Notification, Application.Application.Settings, "uses", $techn="", $tags="", $link="")
+Rel(Application.Application.Notification, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
 Rel(Application.Application.Tenant, Application.Application.Storeprofile, "uses", $techn="", $tags="", $link="")
 Rel(Application.Application.Tenant, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
 Rel(Application.Application.Tenant, Application.Application.Settings, "uses", $techn="", $tags="", $link="")
-Rel(Application.Application.Settings, Application.Application.Shared, "depends on", $techn="", $tags="", $link="")
-Rel(Application.Application.User, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
+Rel(Application.Application.Customer, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
+Rel(Application.Application.Customer, Application.Application.Tenant, "uses", $techn="", $tags="", $link="")
+Rel(Application.Application.Storeprofile, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
 
 SHOW_LEGEND(true)
 @enduml
