+++ backend/docs/modulith/module-customer.puml
@@ -14,9 +14,9 @@ Container_Boundary("Application.Application_boundary", "Application", $tags="") 
   Component(Application.Application.Tenant, "Tenant", $techn="Module", $descr="", $tags="", $link="")
 }
 
+Rel(Application.Application.Tenant, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
 Rel(Application.Application.Customer, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
 Rel(Application.Application.Customer, Application.Application.Tenant, "uses", $techn="", $tags="", $link="")
-Rel(Application.Application.Tenant, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
 
 SHOW_LEGEND(true)
 @enduml
