+++ backend/docs/modulith/module-notification.puml
@@ -15,12 +15,12 @@ Container_Boundary("Application.Application_boundary", "Application", $tags="") 
   Component(Application.Application.Tenant, "Tenant", $techn="Module", $descr="", $tags="", $link="")
 }
 
+Rel(Application.Application.Settings, Application.Application.Shared, "depends on", $techn="", $tags="", $link="")
 Rel(Application.Application.Notification, Application.Application.Tenant, "listens to", $techn="", $tags="", $link="")
 Rel(Application.Application.Notification, Application.Application.Settings, "uses", $techn="", $tags="", $link="")
 Rel(Application.Application.Notification, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
 Rel(Application.Application.Tenant, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
 Rel(Application.Application.Tenant, Application.Application.Settings, "uses", $techn="", $tags="", $link="")
-Rel(Application.Application.Settings, Application.Application.Shared, "depends on", $techn="", $tags="", $link="")
 
 SHOW_LEGEND(true)
 @enduml
