+++ backend/docs/modulith/module-order.puml
@@ -17,17 +17,17 @@ Container_Boundary("Application.Application_boundary", "Application", $tags="") 
   Component(Application.Application.Tenant, "Tenant", $techn="Module", $descr="", $tags="", $link="")
 }
 
-Rel(Application.Application.Customer, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
-Rel(Application.Application.Customer, Application.Application.Tenant, "uses", $techn="", $tags="", $link="")
-Rel(Application.Application.Cast, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
-Rel(Application.Application.Cast, Application.Application.Tenant, "uses", $techn="", $tags="", $link="")
 Rel(Application.Application.Order, Application.Application.Customer, "uses", $techn="", $tags="", $link="")
 Rel(Application.Application.Order, Application.Application.Cast, "uses", $techn="", $tags="", $link="")
 Rel(Application.Application.Order, Application.Application.User, "uses", $techn="", $tags="", $link="")
 Rel(Application.Application.Order, Application.Application.Tenant, "uses", $techn="", $tags="", $link="")
 Rel(Application.Application.Order, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
-Rel(Application.Application.Tenant, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
+Rel(Application.Application.Cast, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
+Rel(Application.Application.Cast, Application.Application.Tenant, "uses", $techn="", $tags="", $link="")
 Rel(Application.Application.User, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
+Rel(Application.Application.Tenant, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
+Rel(Application.Application.Customer, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
+Rel(Application.Application.Customer, Application.Application.Tenant, "uses", $techn="", $tags="", $link="")
 
 SHOW_LEGEND(true)
 @enduml
