+++ backend/docs/modulith/module-tenant.puml
@@ -15,11 +15,11 @@ Container_Boundary("Application.Application_boundary", "Application", $tags="") 
   Component(Application.Application.Tenant, "Tenant", $techn="Module", $descr="", $tags="", $link="")
 }
 
-Rel(Application.Application.Storeprofile, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
+Rel(Application.Application.Settings, Application.Application.Shared, "depends on", $techn="", $tags="", $link="")
 Rel(Application.Application.Tenant, Application.Application.Storeprofile, "uses", $techn="", $tags="", $link="")
 Rel(Application.Application.Tenant, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
 Rel(Application.Application.Tenant, Application.Application.Settings, "uses", $techn="", $tags="", $link="")
-Rel(Application.Application.Settings, Application.Application.Shared, "depends on", $techn="", $tags="", $link="")
+Rel(Application.Application.Storeprofile, Application.Application.Shared, "uses", $techn="", $tags="", $link="")
 
 SHOW_LEGEND(true)
 @enduml
