+++ backend/src/main/java/com/kizuna/user/application/StoreUserService.java
@@ -1,22 +1,43 @@
 package com.kizuna.user.application;
 
+import com.kizuna.shared.exception.ServiceException;
+import com.kizuna.shared.tenancy.TenantScoped;
 import com.kizuna.user.api.dto.StoreUserMeResponse;
+import com.kizuna.user.domain.StoreUser;
+import com.kizuna.user.domain.StoreUserRepository;
+import com.kizuna.user.domain.TenantRole;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-/** 店舗ユーザー自身のアカウント操作サービス。 */
-public interface StoreUserService {
+@Service
+@RequiredArgsConstructor
+public class StoreUserService {
 
-  /**
-   * ログイン中ユーザーの情報を返す。
-   *
-   * @param email 対象ユーザーのメールアドレス（Principal 名）
-   */
-  StoreUserMeResponse me(String email);
+  private final StoreUserRepository userRepository;
 
-  /**
-   * ログイン中ユーザーのニックネームを変更する。
-   *
-   * @param email 対象ユーザーのメールアドレス（Principal 名）
-   * @param nickname 新しいニックネーム
-   */
-  StoreUserMeResponse updateProfile(String email, String nickname);
+  @TenantScoped
+  @Transactional(readOnly = true)
+  public StoreUserMeResponse me(String email) {
+    return toResponse(findByEmail(email));
+  }
+
+  @TenantScoped
+  @Transactional
+  public StoreUserMeResponse updateProfile(String email, String nickname) {
+    StoreUser user = findByEmail(email);
+    user.rename(nickname);
+    return toResponse(userRepository.save(user));
+  }
+
+  private StoreUser findByEmail(String email) {
+    return userRepository
+        .findByEmail(email)
+        .orElseThrow(() -> new ServiceException("ユーザーが見つかりません"));
+  }
+
+  private static StoreUserMeResponse toResponse(StoreUser user) {
+    String role = user.getRoles().stream().findFirst().map(TenantRole::getName).orElse("");
+    return new StoreUserMeResponse(user.getId(), user.getNickname(), user.getEmail(), role);
+  }
 }
